refactor : [007-Scratch-User-02] Test용 스크래치 파일, 사용하지 않는 import 제거
