+++ user-api/src/main/java/com/example/user/controller/CustomerController.java
@@ -8,7 +8,7 @@ import com.example.user.domain.customer.CustomerDto;
 import com.example.user.domain.model.Customer;
 import com.example.user.exception.CustomException;
 import com.example.user.exception.ErrorCode;
-import com.example.user.service.CustomerBalanceService;
+import com.example.user.service.customer.CustomerBalanceService;
 import com.example.user.service.customer.CustomerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
