+++ user-api/src/main/java/com/example/user/service/customer/CustomerBalanceService.java
@@ -1,4 +1,4 @@
-package com.example.user.service;
+package com.example.user.service.customer;
 
 
 import com.example.user.domain.customer.ChangeBalanceForm;
